WEBSEC_MIDTERM/requirements.txt
Flask==3.0.0
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
Pillow>=10.3.0
pytesseract==0.3.10
